componentized delete, added to comments
